Struct tfs::tfsdataframe::TfsDataFrame

pub struct TfsDataFrame { /* fields omitted */ }

The main struct of the crate. TfsDataFrame contains all the information of the loaded TFS file.

TFS file structure

Properties

The first lines are a set of properties of the form

@ PROPERTY_NAME <type-id> value

where <type-id> is one of

type id	type of data

%le	float

%s	string

%d	int

Header

The column header contains of two lines, the first giving the names of the columns and the second one the data types.

The column name row begins with an asterisk (*):

* NAME POSITION col3 ...

The data type row begins with a dollar symbol ($):

$ %s %le ...

Data

And finally, data rows just contain a whitespace separated list of the values for the respective columns.

ELEMENT1 0.0 ...

Methods

impl TfsDataFrame

pub fn open_expect<P>(path: P) -> TfsDataFrame where
P: AsRef<Path>,

Opens a tfs file and stores the conten in a TfsDataFrame. Will panic! if opening fails rather than return a Result<>.~

pub fn open<P>(path: P) -> Result<TfsDataFrame, Error> where
P: AsRef<Path>,

Opens a tfs file and stores the content in a TfsDataFrame.

pub fn set_index(&mut self, column: &str) -> &mut Self

Sets the index to the specified column. The column has to be a String column. f64 does not implement Eq so it cannot be used as hash key.

The index has to be set for loc to work.

use tfs::{DataFrame, TfsDataFrame};

let mut df = TfsDataFrame::open("test/test.tfs").expect("unable to open file");

// works always, indexing using an integer, will acces the nth value according to the order
// in the original file
let betx1 = df.loc_real(0, "BETX").clone();

// only after a first invocation of `set_index` access by the index will be possible
df.set_index("NAME");
let betx2 = df.loc_real("BPM1", "BETX").clone();

assert_eq!(betx1, betx2);

pub fn len(&self) -> usize

pub fn propd(&self, key: &str) -> &f64

Returns the property key from the header if it is a data value, otherwise it panics.

pub fn props(&self, key: &str) -> &String

Returns the property key from the header if it is a string, otherwise it panics.